+++ trunk/source/Library-TraceCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/tracecheckutils/partialorder/independence/ThreadSeparatingIndependenceRelation.java
@@ -24,7 +24,7 @@
  * licensors of the ULTIMATE TraceCheckerUtils Library grant you additional permission
  * to convey the resulting work.
  */
-package de.uni_freiburg.informatik.ultimate.lib.tracecheckutils.independencerelation;
+package de.uni_freiburg.informatik.ultimate.lib.tracecheckutils.partialorder.independence;
 
 import java.util.Objects;
